backend/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.2
flake8==7.1.1
h11==0.14.0
idna==3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.12.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
passlib==1.7.4
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyflakes==3.2.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
ruff==0.7.0
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.40.0
trio==0.27.0
trio-websocket==0.11.1
types-pyasn1==0.6.0.20240913
types-python-jose==3.3.4.20240106
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
